+++ CHANGELOG.md
@@ -31,6 +31,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Fix the bootstrap date icon visualization
 * Update dark mode and accessibility for poll element
 * Add pin element
+* Fix bootstrap validation icon when browser autocomplete function is used
 
 ## [1.4.9] - 05.02.2025
 * Add new "list/list item" element where it's possible to sort the list with checkbox via drag and drop or keyboard